addressed feedback for improved factoids script
